itest: move channel policy update test into one file
